Quick time reload
There is now a box that pops up on the reload timer. if the player pressed F on time the reload will AUTO complete if not the reload will take longer. if the reload is pressed 1-MAX it will act as normal I may change it to any time before half if wanted!