Ammo pickup rework
The ammo pickup works again in the game but I wanted it to be more like a max ammo. This is going to be a base for the pickup but will add a ammo pickup to the inventory system as well, this will let us make mobs drop the pickup!