Sliding Feature
I had to change the player prefab, I had to make the capsule a child of the player entity. Since when I try to scale the capsule while it was the player entity to make it getting lower, the gun will scale down too.

Another thing I changed is the What Is Wall stuff, in the player script I changed it to What To Climb, and the layer to Climbable.

Right now you can slide in any directions, since I don't have a crouch feature yet the player will  get stuck, but I am going to work on that. You just have to move and press the slide key again to start sliding again

I also changed the input manager where Fire 1 can be used by the right control button, Since I assigned the left control button to the sliding feature and upcoming crouch feature.

for some reason for me I couldn't shoot, the gun was always empty. I just pretty sure it's just on my end.