Fixed interaction UI bugs
I made the code a little more incapsulated so different methods arent calling the same bind. instead it checks for the tag pickup and interact then will act differently depending on the tag this leads to the bind only being up when looking at a tag of those, which is what we want.