fixed player sprint bug
There was a bug with the stopping coroutine like a earlier commit but the waitTimer wasnt being stopped sitll so the timer ended up not working. Now both work with getting a private variable first and setting the start to the variable waittime. Then when stopped we set it back to null so there was no null exception.