basic recoil in on the camera.
Now when the guns shoot they will fire in a recoil pattern that was made through lerping/Slerping from one vector to another, resulting in a smooth transition looking like recoil. I made it work in the camera for now since its just rotating the camera but should work fine for our game!